requirements.txt
Flask==3.1.0
flask-cors==5.0.1
PyMySQL==1.1.1